zorko-designer-105 Organize folders according to Docs - more folders move
