--- src/common/utils/createChannelId.ts
@@ -1,3 +1,0 @@
-export const createChannelId = (specId, channelName): string => `${specId}/${channelName}`;
-
-export const extractChannelNameFromId = id => id.split('/')[1];
