+++ src/features/channels/channelsReducerTest.ts
@@ -4,7 +4,7 @@ import * as vegaLiteSpecsFixture from '../__mocks__/vegaLiteSpecsFixtures';
 import {channelsReducer} from './channelsReducer';
 import {chooseSpecFlowReadSuccess} from '../chooseSpecFlow';
 import * as channelsStateFixtures from './__mocks___/channelsStateFixtures';
-import {createChannelId} from '../../common/utils';
+import {createChannelId} from '../../packages/presenterReducerUtils';
 import {encodingChannelFieldEdit} from '../analyticBoard/actions';
 
 describe('Channels Reducer', () => {
