+++ src/features/newEncodings/presenters/EncodingPresenter.ts
@@ -1,4 +1,4 @@
-import {BaseStatePresenter} from '../../../core/BaseStatePresenter';
+import {BaseStatePresenter} from '../../../packages/corePresenters/BaseStatePresenter';
 import {EncodingState} from './EncodingState';
 
 export enum EncodingChannelNames {
