+++ src/features/specs/presenters/SpecsState.ts
@@ -1,4 +1,4 @@
-import {NormalizedState} from '../../../core/normalize/NormalizedState';
+import {NormalizedState} from '../../../packages/corePresenters/normalize/NormalizedState';
 import {SpecState} from './SpecState';
 
 export interface SpecsState extends NormalizedState<SpecState> {}
