+++ src/packages/corePresenters/normalize/BaseNormalizedStateTest.ts
